apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
redis==5.0.8
python-jose[cryptography]==3.3.0
passlib==1.7.4
httpx==0.27.2
email-validator==2.2.0
pytest==8.3.3
ruff==0.6.7
